+++ R/conversion.R
@@ -404,6 +404,35 @@ dim.pandas.core.frame.DataFrame <- function(x) {
     py_object_shape(x)
 }
 
+# Scipy sparse matrices
+
+#' @export
+dim.scipy.sparse.base.spmatrix <- function(x) {
+  if (py_is_null_xptr(x) || !py_available())
+    NULL
+  else
+    py_object_shape(x)
+}
+
+#' @export
+length.scipy.sparse.base.spmatrix <- function(x) {
+  if (py_is_null_xptr(x) || !py_available())
+    0L
+  else
+    prod(py_object_shape(x))
+}
+
+#' @export
+py_to_r.scipy.sparse.base.spmatrix <- function(x) {
+  py_to_r(x$tocsc())
+}
+
+#' @importFrom methods as
+#' @export
+r_to_py.sparseMatrix <- function(x, convert = FALSE) {
+  r_to_py(as(x, "dgCMatrix"), convert = convert)
+}
+
 # Conversion between `Matrix::dgCMatrix` and `scipy.sparse.csc.csc_matrix`.
 # Scipy CSC Matrix: https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csc_matrix.html
 
@@ -419,38 +448,91 @@ r_to_py.dgCMatrix <- function(x, convert = FALSE) {
       x@i, # CSC format index array
       x@p), # CSC format index pointer array
     shape = dim(x))
-  if (any(dim(x) != as_r_value(csc_x$shape)))
+  if (any(dim(x) != dim(csc_x)))
     stop(
-      paste0(
+      paste(
         "Failed to convert: dimensions of the original Matrix::dgCMatrix ",
-        "object and the converted Scipy CSC matrix do not match"))
+        "object (", dim(x), ") and the converted Scipy CSC matrix (",
+        dim(csc_x), ") do not match", sep="", collapse=", "))
   csc_x
 }
 
-#' @importFrom Matrix sparseMatrix
+#' @importFrom methods new
 #' @export
 py_to_r.scipy.sparse.csc.csc_matrix <- function(x) {
   disable_conversion_scope(x)
-  sparseMatrix(
-    i = 1 + as_r_value(x$indices),
-    p = as_r_value(x$indptr),
+  new("dgCMatrix",
+    i = as.integer(as_r_value(x$indices)),
+    p = as.integer(as_r_value(x$indptr)),
     x = as.vector(as_r_value(x$data)),
-    dims = dim(x))
+    Dim = dim(x))
+}
+
+# Conversion between `Matrix::dgRMatrix` and `scipy.sparse.csr.csr_matrix`.
+# Scipy CSR Matrix: https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csr_matrix.html
+
+#' @export
+r_to_py.dgRMatrix <- function(x, convert = FALSE) {
+  # use default implementation if scipy is not available
+  if (!py_module_available("scipy"))
+    return(r_to_py_impl(x, convert = convert))
+  sp <- import("scipy.sparse", convert = FALSE)
+  csr_x <- sp$csr_matrix(
+    tuple(
+      x@x, # Data array of the matrix
+      x@j, # CSR format index array
+      x@p), # CSR format index pointer array
+    shape = dim(x))
+  if (any(dim(x) != dim(csr_x)))
+    stop(
+      paste(
+        "Failed to convert: dimensions of the original Matrix::dgRMatrix ",
+        "object (", dim(x), ") and the converted Scipy CSR matrix (",
+        dim(csr_x), ") do not match", sep="", collapse=", "))
+  csr_x
 }
 
 #' @export
-dim.scipy.sparse.csc.csc_matrix <- function(x) {
-  if (py_is_null_xptr(x) || !py_available())
-    NULL
-  else
-    py_object_shape(x)
+py_to_r.scipy.sparse.csr.csr_matrix <- function(x) {
+  disable_conversion_scope(x)
+  methods::new("dgRMatrix",
+      j = as.integer(as_r_value(x$indices)),
+      p = as.integer(as_r_value(x$indptr)),
+      x = as.vector(as_r_value(x$data)),
+      Dim = dim(x))
 }
+
+# Conversion between `Matrix::dgTMatrix` and `scipy.sparse.coo.coo_matrix`.
+# Scipy COO Matrix: https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.coo_matrix.html
 
 #' @export
-length.scipy.sparse.csc.csc_matrix <- function(x) {
-  if (py_is_null_xptr(x) || !py_available())
-    2L
-  else
-    Reduce(`*`, py_object_shape(x))
+r_to_py.dgTMatrix <- function(x, convert = FALSE) {
+  # use default implementation if scipy is not available
+  if (!py_module_available("scipy"))
+    return(r_to_py_impl(x, convert = convert))
+  sp <- import("scipy.sparse", convert = FALSE)
+  coo_x <- sp$coo_matrix(
+    tuple(
+      x@x, # Data array of the matrix
+      tuple(x@i,
+            x@j)), # COO format coordinate array
+    shape = dim(x))
+  if (any(dim(x) != dim(coo_x)))
+    stop(
+      paste(
+        "Failed to convert: dimensions of the original Matrix::dgTMatrix ",
+        "object (", dim(x), ") and the converted Scipy COO matrix (",
+        dim(coo_x), ") do not match", sep="", collapse=", "))
+  coo_x
 }
 
+#' @importFrom methods new
+#' @export
+py_to_r.scipy.sparse.coo.coo_matrix <- function(x) {
+  disable_conversion_scope(x)
+  new("dgTMatrix",
+      i = as.integer(as_r_value(x$row)),
+      j = as.integer(as_r_value(x$col)),
+      x = as.vector(as_r_value(x$data)),
+      Dim = dim(x))
+}